Add frame rate and frame time trackers
Created a frame_data file that includes structs to hold data for frame
time and functions to calculate and update the current average frame
time and frames per second. Updates once per second.

Could be a good idea to allow for changing the update rate rather than
hardcoding it to be one second.